UA:CMA acutally implement releasing and allocating CMA memory
Userspace can signal whether to allocate or deallocate memory for MFC and FIMC
by writing to /sys/kernel/uacma/enable

0 = disable both,
1 = enable MFC, disable FIMC
2 = enable FIMC,disable MFC
3 = enable both

Add checks to mfc and fimc to not write when CMA is not alloacted.

Also add some dmesg info when loading mfc firmware.
So to make sure dont overwrite it while CMA memory is not allocated.

TODO:
Handle CMA allocation failures.
/sys/kernel/uacma/enable should give status whether the memory is allocated or not.
Like so..
	3:mfc=allocating,fimc=allocated
So on memory pressure we can prompt user to close apps.

Change-Id: I6b7c5e3b46b29e66f3bc146d6b665e5769738593